core: drivers: merge i.MX SNVS driver files
Move the implementation of plat_rpmb_key_is_ready() from
plat-imx/drivers/imx_snvs.c to drivers/imx_snvs.c

Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
