+++ core/arch/arm/plat-imx/drivers/sub.mk
@@ -2,4 +2,3 @@ srcs-$(CFG_TZC380) 	+= tzc380.c
 srcs-$(CFG_IMX_CAAM) 	+= imx_caam.c
 srcs-$(CFG_CSU) 	+= imx_csu.c
 srcs-$(CFG_SCU) 	+= imx_scu.c
-srcs-$(CFG_IMX_SNVS) 	+= imx_snvs.c
